requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
python-dotenv>=1.0.0
openai>=1.12.0
aiohttp>=3.9.3
jinja2>=3.1.3
weasyprint>=64.0
aiosqlite>=0.19.0
biopython>=1.83 